\name{What}
\alias{What do I do?}
\title{What is this?}

\description{
This package is used to display the effects of stimulants when used in a competitive environment. The results indicate that while the user performs better in certain games, they will also experience worse games, whereas the player without the stims is able to maintain a score with significantly less deviation from the average. 
}
